Refactor - Configuration system overhaul
Configuration webpage extended and updated.
Previous settings appear on webpage.
New settings are sent as json str created on client side.
New settings are stored in sd card as json str in text file.
Config is a global dictionary that is updated from sd card.
Configurations are reset and user is forced to configure device if configurations are missing, or sd card was put into another device.
Device unique_id and lora_mac address are displayed on configuration website for user to identify the device.