Add - SHT35 Temperature and Humidity Sensor
The device is now capable of reading, processing and sending data from the high accuracy temperature and humidity sensor from Grove. The temperature sensor is always enabled, because it is needed for PM sensor calibration. The rate at which the sensor takes averages can be changed on the configuration page, but the averaging always takes place at the interval of the PM sensors.

Added Temp_thread.py for reading and logging the temperature and humidity sensor
Added TempSHT35.py for the SHT35 library

Corrected an issue with error logging, where the logger would not be able to print the exception.

Re-added the lora lock, because two lora threads could still run at the same time if the random number generator started the lora thread at the very end of the interval.