Fix configuration button
Previously if the user pressed the button during initialization, the button got reversed, and configuration would start after the user released the button for 2.5 seconds instead of pressing it.
Now the state of the button is known, and user cannot switch polarities, configurations is only entered upon pressing the button for 2.5 seconds.