Add lora send timeout Adjust random interval for sending over LoRa Add check if all keys are in configuration
Usually once lora joined it takes 3 seconds to send the payload. Added a timeout for the socket as well, so now it has a 10 second window to send the payload once it has joined the network.

The lora join timeout is not so reliable it usually takes longer to timeout than the specified interval, therefor 10 seconds leeway was added to the random interval.

All this is very important, because now the lora.csv file is deleted upon sending. If it sends the previous interval's message after calculating averages, it sends the wrong data and an interval worth of data is lost.

Added a developer proof configuration check - now the device forces the user to reconfigure if there is a missing key in the dictionary, previously it was only missing values, now it is both. It is important because when a new default configuration key-value pair is added and it is not configurable by the user it will be missing from the config file since default configurations are only applied when configurations are reset and user is forced to reconfigure.