Add gps init Refactor user button Refactor logging levels Add condition to LoRa
Added gps init function to make gps.py modular so it will be easier to implement getting the location.

User button reboot functionality is initialized at the beginning of main so it can be used to reboot if the sd card failed to initialize. Configuration functionality is only enabled later on when it is appropriate.

Changed some of the level of logging messages because they belong to debug rather than info.

LoRa does not send averages if none of the PM sensors are initialized.

Confirmed, that transistor circuitry works as intended on expansion board v4.

Discovered an interesting behaviour with configurations using expansion board v4 - takes couple of seconds to send configurations - sometimes completely fails.